governance/aws: Add enforce-ami-owners.sentinel for aws_ami data source
This adds enforce-ami-owners.sentinel, which can be used to enforce the
owners list in all aws_ami data sources in a Terraform workspace.

Features:

* Leverages tfconfig, enforcing that the AMI list is actually declared
in configuration, ensuring that values are physically present before any
planning or applies happen.
* Matches declared owners against an expected allow list. Any owner
outside of the allow list results in a failed result.
* Empty owner lists are rejected.
* Resources in violation of the allow list are logged.
* Aggregates results to avoid Sentinel short circuits so that all
violating config instances can be seen at once.
* Displays the full resource address for each violating resource,
properly joined for ease of viewing.

As a consequence of the runtime lacking a join function right now in the
strings import, this policy also contains a join function of its own.
This can be used as an example for other policies that may need it
until we have support for join natively.

For some background on why you would need to enforce owners on aws_ami,
see:
  https://github.com/terraform-providers/terraform-provider-aws/pull/5576